Chapter 8: Metal
Refactor things so that each hittable has a material, which has
different behavior when a ray hits it.

Three materials are introduced: the existing diffuse lambertian material
has moved from the color() function defined in the scene to a material
that is a property of a hittable.  A new metal material bounces a
reflection of the inbound ray, applying its own color via the
attenuation/albedo property.  A debug texture is also introduced that
allows inspection of a hittable's orientation.

A new mechanism that disables anti-aliasing was introduced to speed up
render times during debugging as well as a new SCALE variable that
allows different rendering resolutions to be selected.