Chapter 10: Dielectrics (first half)
Part of chapter 10, which introduces refraction.  Right now, the
refraction is a simplfied model, later in the chapter there will be a
more involved approach to refraction.

I've left a bunch of comments in while trying to understand how the book
arrives at this vector form of snell's law, because it doesn't explain
the derivation.  The derivation in Wikipedia helped and with some
algebra I could eventually make it line up with what the book said to
write.

Also moving to c++.  That might be a result of some other changes that I
started playing around with to paralleize the ray tracker, which I'll
add in a different commit, but I don't mind declaring a dependency on
c++11 now.