Chapter 10: Positionable camera
This was a fun chapter, because the linear algebra was pretty
straightforward and easy for me to reason about.  It was fun guessing
how the scene would look with different transformations and more
intuitive than the refraction calculations from the previous chapter.

Also found some interesting bugs at extreme FOVs (see the last
camera position I added to the scene).  I think that's cased by some
aliasing bug when recasting rays, but I'll have to find that bug another
time.

There's enough here that I can start to build an animation of the scene,
but I'm going to wait to finish the book before adding things like
argument parsing and file I/O that I need to do to make that work well.