Row render scheduling: simplify range math
The -1 and +1's for 'j' are unnecessary and confusing -- they cancel
out.  The results are the same if I just remove them.

I have no idea why I subtracted 1 from ny then added it back in to
every row range calculation for j.  I think I was really tired when I
wrote that.